requirements.txt
amqp==5.3.1
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.2.3
asgiref==3.11.1
billiard==4.2.4
black==26.3.1
boolean.py==5.0
CacheControl==0.14.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.7
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==46.0.7
cyclonedx-python-lib==11.7.0
defusedxml==0.7.1
distlib==0.4.0
Django==6.0.4
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-ipware==7.0.1
django-timezone-field==7.2.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
filelock==3.25.2
firebase_admin==7.3.0
flake8==7.3.0
gitlint==0.19.1
gitlint-core==0.19.1
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
identify==2.6.18
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
license-expression==30.4.4
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.2
mypy_extensions==1.1.0
nodeenv==1.10.0
packageurl-python==0.17.6
packaging==26.0
pathspec==1.0.4
phonenumbers==9.0.27
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==6.33.6
py-serializable==2.1.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
pytest==9.0.2
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-dotenv==1.2.2
python-ipware==3.0.0
pytokens==0.4.1
PyYAML==6.0.3
redis==6.4.0
requests==2.33.1
rich==15.0.0
ruff==0.12.0
sentry-sdk==2.57.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.5
stripe==15.0.1
tomli==2.4.1
tomli_w==1.2.0
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
virtualenv==21.2.1
wcwidth==0.6.0
